[Bronze V] Title: 이상한 기호, Time: 120 ms, Memory: 14236 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,8 +5,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         
         String[] input = (br.readLine()).split(" ");
-        int A = Integer.parseInt(input[0]);
-        int B = Integer.parseInt(input[1]);
+        long A = Long.parseLong(input[0]);
+        long B = Long.parseLong(input[1]);
         
         System.out.print((A + B) * (A - B));
